sidecar/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
aiosqlite>=0.19.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
